kvm: refactor volume mounting
